add -title to be able to set the page's title
