+++ .gitignore
@@ -9,3 +9,4 @@ src/github.com
 # Binary
 /wiki
 /g-wiki
+/g-wiki.exe
