+++ templates/footer.tpl
@@ -1,7 +1,7 @@
 {{define "footer"}}
 <div class="row col-md-9">
   <hr class="text-muted" />
-  <a href="https://github.com/aspic/g-wiki"><p class="text-muted text-center">g-wiki on Github</p></a>
+  <a href="https://github.com/tgulacsi/g-wiki"><p class="text-muted text-center">g-wiki on Github</p></a>
 </div>
 <!-- end container -->
 </div>
