+++ templates/header.tpl
@@ -2,7 +2,7 @@
 <!doctype html>
 <head>
  <meta charset="UTF-8">
- <title>go-wiki!</title>
+ <title>{{.Title}}</title>
  <link href="//netdna.bootstrapcdn.com/bootstrap/3.0.0/css/bootstrap.min.css" rel="stylesheet">
  <meta name="viewport" content="width=device-width, initial-scale=1">
 </head>
